chore: Update eslint rules to disable explicit function return type and prettier
This commit updates the `.eslintrc.cjs` file to disable the rule `@typescript-eslint/explicit-function-return-type` and `prettier/prettier`. This change is made to align with the recent updates in npm dependencies and build scripts, which included updates to dev dependencies such as `@electron-toolkit/eslint-config-ts` and `eslint`. By disabling these rules, the codebase will no longer enforce explicit function return types and prettier formatting. This commit is part of the ongoing maintenance and configuration updates for the project.