refactor: Update Home component to add user creation functionality
This commit updates the Home component in the src/renderer/src/Home.tsx file to include the functionality for creating users. It adds a button that triggers the handleAddUser function, which makes an API call to create a new user. After the user is created, the getUsers query is invalidated to fetch the updated list of users. This change enhances the Home component by allowing users to be added dynamically.