refactor: Add getUsers and createUser queries to greetingRouter
This commit adds the getUsers and createUser queries to the greetingRouter in the src/main/routers/greeting.ts file. The getUsers query fetches the list of users from the database using the db.select() method, while the createUser query inserts a new user into the database using the db.insert() method. These changes enhance the functionality of the greetingRouter by allowing users to be fetched and created.