feat: Add script to start local development database container
This commit adds a new script, `start-database.sh`, to the repository. The script is used to start a docker container for a local development database. It includes instructions for running the script on different operating systems and checks for the presence of Docker. The script also handles the creation of a new database container if it doesn't already exist, and generates a random password if the default password is being used. This change improves the development workflow by providing a convenient way to start the database container.