feat: Add users table to database schema
This commit adds a new table called "users" to the database schema. The table includes columns for "id" and "textModifiers". The "textModifiers" column is set to be not null and has a default value of the current timestamp. This change enhances the database structure by providing a table to store user information.